+++ src/lib/builders/collapsible/tests/CollapsibleTest.svelte
@@ -1,0 +1,14 @@
+<script lang="ts">
+	import { createCollapsible } from '..';
+
+	export let open = false;
+
+	const { root, trigger, content } = createCollapsible({
+		open,
+	});
+</script>
+
+<div {...$root}>
+	<button {...$trigger} data-testid="trigger">Trigger</button>
+	<div {...$content} data-testid="content">Content</div>
+</div>