+++ src/lib/builders/index.ts
@@ -4,3 +4,8 @@ export * from './popover';
 export * from './select';
 export * from './tabs';
 export * from './toggle';
+export * from './checkbox';
+export * from './switch';
+export * from './toggle-group';
+export * from './progress';
+export * from './radio-group';