+++ src/routes/docs/builders/collapsible/+page.svelte
@@ -11,4 +11,4 @@
 
 <Docs.CodeBlock {code} />
 
-<!-- <Docs.H2>API Reference</Docs.H2> -->
+<Docs.Construction />