+++ src/routes/docs/builders/switch/+page.svelte
@@ -1,0 +1,16 @@
+<script lang="ts">
+	import { Docs } from '$routes/(components)';
+	import code from './code.ignore-svelte?raw';
+	import Preview from './preview.svelte';
+</script>
+
+<Docs.H1>Switch</Docs.H1>
+<Docs.Description>
+	A control that allows the user to toggle between checked and not checked.
+</Docs.Description>
+
+<Preview />
+
+<Docs.CodeBlock {code} />
+
+<Docs.Construction />