+++ src/routes/docs/builders/tabs/+page.svelte
@@ -12,3 +12,5 @@
 <Preview />
 
 <Docs.CodeBlock {code} />
+
+<Docs.Construction />