+++ src/routes/docs/builders/toggle/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { Docs } from '$routes/(components)';
-	import code from './code.svelte?raw';
+	import code from './code.ignore-svelte?raw';
 	import Preview from './preview.svelte';
 </script>
 
@@ -11,4 +11,4 @@
 
 <Docs.CodeBlock {code} />
 
-<!-- <Docs.H2>API Reference</Docs.H2> -->
+<Docs.Construction />