+++ tailwind.config.ts
@@ -87,8 +87,18 @@ export default {
 	},
 
 	plugins: [
-		plugin(function ({ addVariant }) {
+		plugin(function ({ addVariant, matchUtilities, theme }) {
 			addVariant('hocus', ['&:hover', '&:focus']);
+			// Square utility
+			matchUtilities(
+				{
+					square: (value) => ({
+						width: value,
+						height: value,
+					}),
+				},
+				{ values: theme('spacing') }
+			);
 		}),
 	],
 } satisfies Config;