+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # @melt-ui/svelte
 
+## 0.3.0
+
+### Minor Changes
+
+- 1ba6b49: add radio group
+- 5c4ed8c: add progress
+
+## 0.2.0
+
+### Minor Changes
+
+- 2377592: add switch
+- b685910: add toggle group
+
 ## 0.1.0
 
 ### Minor Changes